+++ sov-modules/sov-modules-api/src/lib.rs
@@ -1,19 +1,21 @@
 #![feature(associated_type_defaults)]
+#[cfg(feature = "mocks")]
+pub mod mocks;
 
 mod encode;
 mod error;
-#[cfg(feature = "mocks")]
-pub mod mocks;
 mod prefix;
+mod response;
 
 pub use error::{DecodingError, DispatchError, ModuleError};
 pub use prefix::Prefix;
+pub use response::{CallResponse, QueryResponse};
+
 use sov_state::Storage;
-use sovereign_sdk::{
-    serial::{Decode, Encode},
-    stf::{Event, EventKey, EventValue},
-};
-use std::{fmt::Debug, rc::Rc};
+use sovereign_sdk::serial::{Decode, Encode};
+
+/// A type that can't be instantiated.
+pub enum NonInstantiable {}
 
 /// Context contains types and functionality common for all modules.
 pub trait Context {
@@ -23,33 +25,6 @@ pub trait Context {
 
     /// Sender of the transaction.
     fn sender(&self) -> &Self::PublicKey;
-}
-
-/// A type that can't be instantiated.
-pub enum NonInstantiable {}
-
-/// Response type for the `Module::call` method.
-#[derive(Default)]
-pub struct CallResponse {
-    /// Lists of events emitted by a call to a module.
-    events: Vec<Event>,
-}
-
-impl CallResponse {
-    pub fn add_event(&mut self, key: &str, value: &str) {
-        let event = Event {
-            key: EventKey(Rc::new(key.as_bytes().to_vec())),
-            value: EventValue(Rc::new(value.as_bytes().to_vec())),
-        };
-
-        self.events.push(event)
-    }
-}
-
-/// Response type for the `Module::query` method. The response is returned by the relevant RPC call.
-#[derive(Default, Debug)]
-pub struct QueryResponse {
-    pub response: Vec<u8>,
 }
 
 /// Every module has to implement this trait.
